[deploy] Update core containers for Git FS.
This commit contains multiple changes.

First, the core containers have been bumped to RockyLinux 10 to access
the packaged version of python3.12.

Second, the initial DB setup script has been updated to load an initial
SSH private key for the superadmin. This key is used to authenticate on
the gitolite instance.
The frontend service image has been updated to leverage this extension
of the DB setup script on service startup.

Third, previous commits leverage a patched version of pygit2. While the
patch https://github.com/libgit2/pygit2/pull/1395 is not available
through PyPi, we have to leverage a custom build of pygit2. However, the
packaged libgit2 version on RL10 is too old for current versions of
pygit2, hence we also have to custom build libgit2 and libssh2.