[frontend/fs/git] Push course to remote upon creation.
Upon course creation, we ensure that the course and its $common submodule
are initialized on the remote through gitolite configuration update. We
then locally create the course and the $common submodule repositories.
They contains the descriptors defined through the web interface. The
course and its $common submodule are then pushed on the remote.

This requires storing a private SSH key for users and propagating credential
information within the Git FSP to ensure commits authorship and remote
authentication.

This commit also improves the gitolite interface tests and adds specific
tests for the Git FSP.